Extract admin warning from class-parsely (#468)
* Moving settings page out of class-parsely

* Move print dynamic tracking note

* Fixing hooks

* Fix automated linting issues

* Fix manual linting issues

* Extracting admin warning

* Updating hook to _admin_menu

* Moving admin warning tests

* Fix get method
