+++ wp-parsely.php
@@ -30,6 +30,7 @@ use Parsely;
 use Parsely\Integrations\Amp;
 use Parsely\Integrations\Facebook_Instant_Articles;
 use Parsely\Integrations\Integrations;
+use Parsely\UI\Admin_Warning;
 use Parsely\UI\Plugins_Actions;
 use Parsely\UI\Recommended_Widget;
 use Parsely\UI\Row_Actions;
@@ -52,13 +53,16 @@ add_action(
 );
 
 // Until auto-loading happens, we need to include this file for tests as well.
+require __DIR__ . '/src/UI/class-admin-warning.php';
 require __DIR__ . '/src/UI/class-plugins-actions.php';
 require __DIR__ . '/src/UI/class-row-actions.php';
 add_action(
 	'admin_init',
 	function(): void {
-		$GLOBALS['parsely_ui_plugins_actions'] = new Plugins_Actions();
+		$admin_warning = new Admin_Warning( $GLOBALS['parsely'] );
+		$admin_warning->run();
 
+		$GLOBALS['parsely_ui_plugins_actions'] = new Plugins_Actions();
 		$GLOBALS['parsely_ui_plugins_actions']->run();
 
 		$row_actions = new Row_Actions( $GLOBALS['parsely'] );
